// convex/_generated/api.d.ts
declare const fullApi: ApiFromModules<{
  appointments: typeof appointments;
  auth: typeof auth;
  cases: typeof cases;
  clients: typeof clients;
  dashboard: typeof dashboard;
  http: typeof http;
  model: typeof model;
  seed: typeof seed;
  settings: typeof settings;
  treatments: typeof treatments;
  users: typeof users;
}>;

/**
 * A utility for referencing Convex functions in your app's public API.
 *
 * Usage:
 * ```js
 * const myFunctionReference = api.myModule.myFunction;
 * ```
 */
export declare const api: FilterApi<
  typeof fullApi,
  FunctionReference<any, "public">
>;